ui (fix): update from dpreid fork
timdrysdale moved the ui to vite for use with the new system, so this
commit is combining the vite build files from that effort with the
latest source from dpreid's forked version of the repo. Migrating this
ui involved updating and re-patching our fork of the jsmpeg-player dependency to
avoid a node gyp error on trying to install it.